portfolio_app_v3/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
sqlalchemy>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.40
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
